Add key / argument form
Added form for calls. This gathers a list of argument fields based on
the names listed in the database, including two for the key, if it takes
one. postCall runs the form and gets its result, but doesn't do anything
with it yet. Most field names are using a "not implemented" placeholder.

There is some parsing in the fields that have been implemented, but none
of it has been tested yet. Many of the helper functions are probably
useful enough elsewhere to be lifted to the top level.

Other changes:
* Factored out the page-rendering parts of getCall and postCall.
* Added Attoparsec to dependencies.
* Added unknown key scope, for use between receiving a key and checking
  its properties.
* Added keyOpt record field label to non-empty KeyType constructors for
  convenience.
* Added FullCall type containing everything needed to make a call (key,
  args, etc).
* Removed keyOptional field from Key (it's for spec types, not concrete
  data types).
* readArg proved useless, so was removed.