Implement CharacterID and CorporationID
As a side effect, the key fields are now explicitly processed before the
argument fields, since we need to know whether they're present for
FromKey arguments. Also the final result expression is a big uglier due
to the necessity of monadic code. On the plus side, there are more
comments now.