Add tree view magic
A javascript file for presenting unordered lists in a collapsible tree
view a la the left pane of Windows Explorer. To use, put the LinkedList
class on the root ul of any lists you want to collapse, and put
addEvents in the document onload hook.

(This should be redundant with HTML 5's details element, but only Chrome
implements it.)

TODO: use less ugly expanded/collapsed icons.