Refactor methods
Moved the generic make-the-call code out of callList so other calls can
have trivial implementations, and added a function to check that the
parameters are correct and normalise them. This entails passing in the
call parameters as described in the database along with the HTTP
manager.

All calls that take no key and no arguments are now implemented.

For avoidance of confusion, CallArgs (which includes the key if any) was
renamed CallParams.

Additionally, API calls are now made using POST.